+++ CHANGELOG.md
@@ -43,6 +43,8 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
   * updated business logic to consume use case data due to updated api response body style
 * Service Marketplace - Service Detail Page
   * updated business logic to consume use case data due to updated api response body style
+* App Approval Release
+   * Added missing long description in app details
 
 ## 1.2.0